test: Move timeline edit integration test into new module
